app/requirements.txt
uvicorn
fastapi
python-dotenv
sqlalchemy
python-dateutil